requirements.txt
absl-py==2.1.0
accelerate==1.1.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.44.1
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
chainer==7.8.1
charset-normalizer==3.4.0
clean-fid==0.1.35
click==8.1.7
colorama==0.4.6
colossalai==0.4.6
contexttimer==0.3.3
contourpy==1.3.0
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
decorator==5.1.1
decord==0.6.0
Deprecated==1.2.14
diffusers==0.32.2
distlib==0.3.9
docstring_parser==0.16
easydict==1.13
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fabric==3.2.2
facexlib==0.3.0
fairscale==0.4.13
fastapi==0.115.4
ffmpy==0.5.0
filelock==3.16.1
filterpy==1.4.5
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
galore-torch==1.0
google==3.0.0
gradio==5.12.0
gradio_client==1.5.4
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.1
icecream==2.1.3
identify==2.6.1
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
iniconfig==2.0.0
invoke==2.2.0
iopath==0.1.10
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.8
llvmlite==0.43.0
lmdb==1.5.1
loguru==0.7.2
lpips==0.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.2
memory-profiler==0.61.0
mmengine==0.10.5
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
ninja==1.11.1.1
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.17.0
openai-clip==1.0.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opencv-transforms==0.0.6
orjson==3.10.15
oss2==2.19.1
packaging==24.1
pandas==2.2.3
paramiko==3.5.0
parso==0.8.4
peft==0.13.2
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
plumbum==1.9.0
portalocker==3.0.0
pre_commit==4.0.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyiqa==0.1.13
PyNaCl==1.5.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-fid==0.3.0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
ray==2.38.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.4
rotary-embedding-torch==0.8.4
rpds-py==0.20.1
rpyc==6.0.0
ruff==0.9.2
safehttpx==0.1.6
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.11
tokenizers==0.15.2
tomli==2.0.2
tomlkit==0.13.2
torch==2.3.0
torch-dist @ http://visionai-wlcb.oss-cn-wulanchabu.aliyuncs.com/library/whl/torch_dist-1.2.9-py3-none-any.whl#sha256=92ae7dfce3b1adf1c02f041394c382aadc649247f397561db048a111ed7d7a81
torch-fidelity==0.3.0
torchaudio==2.3.0
torchmetrics==1.5.1
torchvision==0.18.0
tqdm==4.66.5
traitlets==5.14.3
transformers==4.37.2
triton==2.3.0
typed-argument-parser==1.10.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uuid==1.30
uvicorn==0.29.0
virtualenv==20.27.1
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.1.3
wrapt==1.16.0
xformers==0.0.26.post1
yacs==0.1.8
yapf==0.40.2
yarl==1.17.1
zipp==3.20.2
